py3/requirements.txt
boto3==1.12.36
botocore==1.15.36
click==7.1.1
colorama==0.4.3
docutils==0.15.2
jmespath==0.9.5
numpy==1.18.3
python-dateutil==2.8.1
PyYAML==5.3.1
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.8
yapf==0.29.0
